% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_CIT.R
\name{cit.cox}
\alias{cit.cox}
\title{Causal Inference Test for survival outcomes}
\usage{
cit.cox = function( L, G, T, E, CT=NULL, n.resampl=50, n.perm=0, perm.index=NULL, rseed=NULL )
}
\arguments{
\item{L}{Vector or nxp design matrix representing the instrumental variable(s).}

\item{G}{Continuous vector representing the potential causal mediator.}

\item{T}{Continuous vector representing time values for time-to-event outcome.}

\item{E}{Vector of events, 0:censored, 1:event.}

\item{CT}{Vector or nxp design matrix representing adjustment covariates when Y is T.}

\item{n.resampl}{The number of instances of the test statistic for conditional independence (test 4) generated by ermutation (Millstein et al. 2009) under the null hypothesis of no mediation (independent effects of L on G and T). These data are used to estimate the parameters of the null distribution. The default is set to 50, which we have found to provide reasonable precision.}

\item{n.perm}{Number of permutations for each component test if greater than 0.}

\item{perm.index}{An n x n.perm matrix of permutation indices.}

\item{rseed}{Seed for reproducible permutations.}
}
\value{
A dataframe which includes the following columns:
\item{perm}{Indicator for permutation results. Zero indicates that the data were not permuted and subsequent rows include an integer greater than zero for each permutation conducted.}
\item{p_cit}{CIT (omnibus) p-value}
\item{p_TassocL}{component p-value for the test of association between T and L.}
\item{p_TassocGgvnL}{component p-value for the test of association between T and G|L.}
\item{p_LindTgvnG}{component p-value for the equivalence test of L ind T|G}
}
\description{
This function implements a formal statistical hypothesis test, resulting in a p-value, to quantify uncertainty in a causal inference pertaining to a measured factor, e.g. a molecular species, which potentially mediates a known causal association between a locus or other instrumental variable and a quantitative trait. If the number of permutations is greater than zero,  then the results can be used with fdr.cit to generate permutation-based FDR values (q-values) that are returned with confidence intervals to quantify uncertainty in the estimate. The outcome is continuous, the potential mediator is continuous, and the instrumental variable can be continuous, discrete (such as coding a SNP 0, 1, 2), or binary and is not limited to a single variable but may be a design matrix representing multiple variables.
}
\details{
Increasing n.resampl will increase the precision of the component test 4, the conditional independence test. This may be useful if a very small p-value is observed and high precision is desired, cit.cp 7 however, it will increase run time. The omnibus p-value, p_cit, is the maximum of the component p-values, an intersection-union test, representing the probability of the data if at least one of the component null hypotheses is true. If permutations are conducted by setting n.perm to a value greater than zero, then the results are provided in matrix (dataframe) form, where each row represents an analysis using a unique permutation, except the first row (perm = 0), which has results from the observed or non-permuted analysis. These results can then be aggregated across multiple cit.cp tests and input to the function fdr.cit to generate component test FDR values (q-values) as well as omnibus q-values with confidence intervals that correspond to the p_cit omnibus p-values.
}
\examples{
# Sample Size
ss = 100

# Errors
e1 = matrix(rnorm(ss * 3), ncol=3)
e2 = matrix(rnorm(ss * 3), ncol=3)

# Simulate genotypes, gene expression, covariates, and clinical trait matrices
L = matrix(rbinom(ss * 3, 2, 0.5), ncol=3)
LS = matrix(rowSums(L), ncol = 1)
e1s = matrix(rowSums(e1), ncol = 1)
G = matrix(apply(0.3 * LS , 1, sum) + e1s, ncol=1)
E = rbinom(ss, 1, 0.5)
T = rexp(ss, rate = 0.1)
CT = matrix(matrix(rnorm(ss*2), ncol=1), ncol=2)

n.perm = 5
perm.index = matrix(NA, nrow=ss, ncol=n.perm)
for(j in 1:ncol(perm.index)) perm.index[, j] = sample(1:ss)

# Run tests
results = cit.cox(L, G, T, E)
results

results = cit.cox(L, G, T, E, perm.index=perm.index, n.perm=5)
results

results = cit.cox( L, G, T, E, CT)
results

results = cit.cox( L, G, T, E, CT, n.perm=5)
results

}
\references{
Millstein J, Chen GK, Breton CV. 2016. cit: hypothesis testing software for mediation analysis in genomic applications. Bioinformatics. PMID: 27153715.
Millstein J, Zhang B, Zhu J, Schadt EE. 2009. Disentangling molecular relationships with a causal inference test. BMC Genetics, 10:23.
}
\author{
Joshua Millstein, Mingzhi Ye
}
